Finished failed request tracing.

+++ JexusManager.Shared/Features/TimeSpanExtensions.cs
@@ -13,7 +13,7 @@ namespace JexusManager.Features
             return time.Minutes + (time.Hours + time.Days * 24) * 60;
         }
 
-        public static int GetTotalSeconds(this TimeSpan time)
+        public static long GetTotalSeconds(this TimeSpan time)
         {
             return time.Seconds + (time.Minutes + (time.Hours + time.Days * 24) * 60) * 60;
         }
